removes redundant LevelState declaration from session module declaration
