+++ backend/test/integration/chatController.test.ts
@@ -16,12 +16,6 @@ declare module 'express-session' {
 		chatModel: ChatModel;
 		levelState: LevelState[];
 	}
-	interface LevelState {
-		level: LEVEL_NAMES;
-		chatHistory: ChatMessage[];
-		defences: Defence[];
-		sentEmails: EmailInfo[];
-	}
 }
 
 // mock the api call
