+++ backend/test/unit/controller/emailController.test.ts
@@ -4,22 +4,14 @@ import { Response } from 'express';
 import { handleClearEmails } from '@src/controller/emailController';
 import { EmailClearRequest } from '@src/models/api/EmailClearRequest';
 import { ChatModel } from '@src/models/chat';
-import { ChatMessage } from '@src/models/chatMessage';
-import { Defence } from '@src/models/defence';
 import { EmailInfo } from '@src/models/email';
-import { LEVEL_NAMES } from '@src/models/level';
+import { LevelState } from '@src/models/level';
 
 declare module 'express-session' {
 	interface Session {
 		initialised: boolean;
 		chatModel: ChatModel;
 		levelState: LevelState[];
-	}
-	interface LevelState {
-		level: LEVEL_NAMES;
-		chatHistory: ChatMessage[];
-		defences: Defence[];
-		sentEmails: EmailInfo[];
 	}
 }
