+++ backend/test/unit/controller/resetController.test.ts
@@ -19,13 +19,8 @@ declare module 'express-session' {
 		chatModel: ChatModel;
 		levelState: LevelState[];
 	}
-	interface LevelState {
-		level: LEVEL_NAMES;
-		chatHistory: ChatMessage[];
-		defences: Defence[];
-		sentEmails: EmailInfo[];
-	}
 }
+
 function responseMock() {
 	return {
 		send: jest.fn(),
